requirement.txt
asgiref==3.8.1
Django==4.2.16
django-environ==0.11.2
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
